+++ src/main.ts
@@ -1,5 +1,5 @@
 import { createApp } from 'vue'
-import { Notify, Quasar } from 'quasar'
+import { LoadingBar, Notify, Quasar } from 'quasar'
 import { createPinia } from 'pinia'
 import VueApexCharts from 'vue3-apexcharts'
 
@@ -11,7 +11,14 @@ import router from './router/index'
 
 createApp(App)
   .use(Quasar, {
+    config: {
+      loadingBar: {
+        color: 'amber',
+        size: '7px',
+      },
+    },
     plugins: {
+      LoadingBar,
       Notify,
     },
   })